Updated a few functions here. First, i fixed a call to uniprot that attempted to convert from KEGG to GENENAME. Uniprot doesn't allow this so I had to go from KEGG to uniprot ACC to GENENAME. Stupid but I see why they did it. Uniprot ACC can be unreviewed. So I then had to find all reviewed terms. SIlly but it works for now.
In the long run we should try to download all of kegg and convert to uniprot right off the back. Then we replace the code that pulls down each KGML and converts it to imported an already converted file. Then we can have a function to update all the KGML files.  This will save a lot of the time creating networks, which is spent communications with the REST apis.