Updating visualization tools. Improved the cytoscape.js styles. Now things can be modified more easily using python, not going through json files.
Updating network functions just for igraph. Can now just pass a any network and it generates a figure.

Added an enrichment tools that slims the ontology per time point (or sample). Before I did it on an entire array but after careful consideration, and testing, i find that it might remove too many things depending on when it showed up compared so something with a higher score.